ospfd: OSPF redistributes /32 from loopback regardless of prefix length
This commit removes the hard-coded /32 limitation on connected networks
on the loopback interface.

Setting a secondary IP, e.g., 10.0.0.1/24 on interface "lo" for
redistribution as a connected network, results in only injecting
10.0.0.1/32 LSA to neigbors.

There are several uses for the loopback interface, e.g., when you want
announce a network that is to be 1-to-1 NATed.  In any case, it should
not be up to OSPF to limit what the user wants to do in this case.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>